+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### Unreleased
 
+* Migrate `sting-server` transaction support from `javax.transaction` to `jakarta.transaction-api` 2.0.1.
+
 ### [v0.38](https://github.com/sting-ioc/sting/tree/v0.38) (2026-06-23) · [Full Changelog](https://github.com/sting-ioc/sting/compare/v0.37...v0.38) · [API Differences](https://sting-ioc.github.io/api-diff?key=sting&old=0.37&new=0.38)
 
 The release includes 13 non breaking API changes.